[Update] SVN Cleanup and Fixed warnings
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk/MMOCoreORB@1505 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ src/server/zone/managers/feature/FeatureManager.h
@@ -82,7 +82,7 @@ public:
 	}
 
 	inline int getFloatFeature(const String& key) {
-		return floatFeature->get(key);
+		return (int)floatFeature->get(key);
 	}
 
 	inline bool hasSetFeature(const String& key, const String& feature) {
